Added Class Rational: which explains using operators as method names and implicit

+++ src/main/scala/org/impatient/ch8/Item.scala
@@ -19,7 +19,7 @@ class Bundle(val description: String) extends Item {
     item :: items
   }
 
-  override def price(): Double = {
+  override def price(): Int = {
     items.map(_.price()).sum
   }
 }
